Use accurate algorithm for all predictors
